+++ src/Fable.Transforms/Python/Fable2Python.fs
@@ -2307,7 +2307,14 @@ module Util =
         let expr, stmts = com.TransformAsExpr(ctx, guardExpr)
 
         match expr with
-        | Constant(value = BoolLiteral value) -> stmts @ com.TransformAsStatements(ctx, ret, thenStmnt)
+        | Constant(value = BoolLiteral value) ->
+            let e =
+                if value then
+                    thenStmnt
+                else
+                    elseStmnt
+
+            stmts @ com.TransformAsStatements(ctx, ret, e)
         | guardExpr ->
             let thenStmnt, stmts' =
                 transformBlock com ctx ret thenStmnt
@@ -4143,8 +4150,14 @@ module Util =
                 if List.isEmpty interfaces then
                     com.GetImportExpr(ctx, "typing", "Protocol")
 
-                for gen in classEnt.GenericParameters do
-                    Expression.subscript (com.GetImportExpr(ctx, "typing", "Generic"), com.AddTypeVar(ctx, gen.Name))
+                // Collect all generic type variables and create a single `Generic` base
+                let genericTypeVars =
+                    classEnt.GenericParameters
+                    |> Seq.map (fun gen -> com.AddTypeVar(ctx, gen.Name))
+                    |> Seq.toList
+
+                if not (List.isEmpty genericTypeVars) then
+                    Expression.subscript (com.GetImportExpr(ctx, "typing", "Generic"), Expression.tuple genericTypeVars)
             ]
 
         [ Statement.classDef (classIdent, body = classMembers, bases = bases) ]
@@ -4288,7 +4301,8 @@ module Util =
         // printfn "getIdentForImport: %A" (moduleName, name)
         match name with
         | None -> Path.GetFileNameWithoutExtension(moduleName)
-        | Some name -> name |> Naming.toSnakeCase |> getUniqueNameInRootScope ctx
+        | Some name -> name |> Naming.toSnakeCase
+        |> getUniqueNameInRootScope ctx
         |> Identifier
 
 module Compiler =