+++ src/Fable.Cli/CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+
+* [Python] Fixed testing with interfaces leads to incorrect pattern match (#3972) (by @dbrattli)
+* [Python] Fixed error when type contains multiple generic type parameters (#3986) (by @dbrattli)
+* [Python] Fixed import path handling for libraries (#4088) (by @dbrattli)
+* [Python] Reenable type aliasing for imports with name "*" (by @freymauer)
+
 ### Removed
 
 * [All] Remove unused `Fable.Core` copy types from internal module (by @MangelMaxime)