+++ src/Fable.Cli/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## 5.0.0-alpha.13 - 2025-05-04
+
 ### Fixed
 
 * [Python] Fixed testing with interfaces leads to incorrect pattern match (#3972) (by @dbrattli)